requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-wtf==1.1.1
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
email-validator==2.0.0
WTForms==3.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
alembic==1.11.3
SQLAlchemy==2.0.20